begin creating the dao for book
