+++ app/dao/__init__.py
@@ -1,0 +1,1 @@
+from dao import BookDao
